chore: separate discussions logic
