+++ lib/agenda.js
@@ -1,5 +1,7 @@
 'use strict'
 
+const { graphql } = require('@octokit/graphql')
+
 /**
  * get agenda issues and PRs from repositories
  * @param {Object} client - GitHub client
@@ -38,7 +40,7 @@ async function fetchAgendaItems (client, repos, agendaLabel) {
       labels: agendaLabel,
       per_page: 100
     })).filter(pr => pr.labels.find(label => label.name === agendaLabel) &&
-                     !(agendaIssues.find((i) => i.url === pr.url))) // workaround for flaky GH API/SDK behavior where sometimes the issue endpoint loads PRs
+      !(agendaIssues.find((i) => i.url === pr.url))) // workaround for flaky GH API/SDK behavior where sometimes the issue endpoint loads PRs
 
     console.log(`Fetching PRs for ${r.owner}/${r.repo}: Found ${_agendaPrs.length}`)
 
@@ -52,6 +54,80 @@ async function fetchAgendaItems (client, repos, agendaLabel) {
   return agendaIssues
 }
 
+async function fetchDiscussionsItems (repos, agendaLabel, token) {
+  const agendaDiscussions = []
+  for (const r of repos) {
+    let hasNextPage = true
+    let endCursor = null
+    do {
+      const query = `
+          query($owner: String!, $name: String!, $after: String) {
+            repository(owner: $owner, name: $name) {
+              discussions(first: 100, after: $after) {
+                pageInfo {
+                  endCursor
+                  hasNextPage
+                }
+                edges {
+                  cursor
+                  node {
+                    id
+                    title
+                    url
+                    labels(first: 10) {
+                      nodes {
+                        color
+                        name
+                      }
+                    }
+                  }
+                }
+              }
+            }
+          }
+        `
+      const variables = {
+        owner: r.owner,
+        name: r.repo,
+        after: endCursor
+      }
+
+      const _agendaDiscussions = await graphql(query, {
+        ...variables,
+        headers: {
+          authorization: `token ${token}`
+        }
+      })
+
+      const discussions = _agendaDiscussions?.repository?.discussions
+
+      if (discussions) {
+        const { edges, pageInfo } = discussions
+        for (const edge of edges) {
+          const labels = edge.node?.labels.nodes
+          if (Array.isArray(labels) && labels.some(label => label.name === agendaLabel)) {
+            console.log(`Adding Discussion: ${edge.node.url}`)
+            agendaDiscussions.push({
+              id: edge.node.id,
+              html_url: edge.node.url,
+              title: edge.node.title
+            })
+          }
+        }
+        hasNextPage = pageInfo.hasNextPage
+        endCursor = pageInfo.endCursor
+      } else {
+        hasNextPage = false
+      }
+    } while (hasNextPage)
+  }
+
+  console.log(`Found ${agendaDiscussions.length} total discussions for agenda`)
+
+  return agendaDiscussions
+}
+
 module.exports = {
-  fetchAgendaItems
+  fetchAgendaItems,
+  fetchDiscussionsItems
 }
