+++ package.json
@@ -27,7 +27,7 @@
     "@actions/github": "^6.0.1",
     "@hackmd/api": "^2.5.0",
     "@js-temporal/polyfill": "^0.5.1",
-    "@octokit/graphql": "^9.0.1",
+    "@octokit/graphql": "^8.2.2",
     "ejs": "^3.1.10",
     "safe-parse-list": "^0.1.1"
   },
