+++ run.js
@@ -1,7 +1,6 @@
 'use strict'
 const core = require('@actions/core')
 const { getOctokit, context } = require('@actions/github')
-const { graphql } = require('@octokit/graphql')
 const list = require('safe-parse-list')
 const ejs = require('ejs')
 const meetings = require('./lib/meetings')
@@ -31,6 +30,8 @@ const pkg = require('./package.json')
     // variables we use for content
     const issueTitle = core.getInput('issueTitle')
     const issueTemplate = core.getInput('issueTemplate')
+
+    const discussions = core.getBooleanInput('discussions')
 
     // variables we use for notes
     const createNotes = core.getInput('createNotes')
@@ -97,72 +98,15 @@ const pkg = require('./package.json')
       }
     }
 
-    const agendaIssues = await agenda.fetchAgendaItems(client, repos, agendaLabel)
-  
-    for (const r of repos) {
-      let hasNextPage = true
-      let endCursor = null
-      do {
-        const query = `
-          query($owner: String!, $name: String!, $after: String) {
-            repository(owner: $owner, name: $name) {
-              discussions(first: 100, after: $after) {
-                pageInfo {
-                  endCursor
-                  hasNextPage
-                }
-                edges {
-                  cursor
-                  node {
-                    id
-                    title
-                    url
-                    labels(first: 10) {
-                      nodes {
-                        color
-                        name
-                      }
-                    }
-                  }
-                }
-              }
-            }
-          }
-        `
-        const variables = {
-          owner: r.owner,
-          name: r.repo,
-          after: endCursor
-        }
-        const _agendaDiscussions = await graphql(query, {
-          ...variables,
-          headers: {
-            authorization: `token ${token}`
-          }
-        })
-        const discussions = _agendaDiscussions?.repository?.discussions
-        if (discussions) {
-          const { edges, pageInfo } = discussions
-          for (const edge of edges) {
-            const labels = edge.node?.labels.nodes
-            if (Array.isArray(labels) && labels.some(label => label.name === agendaLabel)) {
-              console.log(`Adding Discussion: ${edge.node.url}`)
-              agendaIssues.push({
-                id: edge.node.id,
-                html_url: edge.node.url,
-                title: edge.node.title
-              })
-            }
-          }
-          hasNextPage = pageInfo.hasNextPage
-          endCursor = pageInfo.endCursor
-        } else {
-          hasNextPage = false
-        }
-      } while (hasNextPage)
+    const agendaItems = []
+    agendaItems.push(...await agenda.fetchAgendaItems(client, repos, agendaLabel))
+
+    if (discussions) {
+      const agendaDiscussions = await agenda.fetchDiscussionsItems(repos, agendaLabel, token)
+      agendaItems.push(...agendaDiscussions)
     }
 
-    console.log(`Found ${agendaIssues.length} total issues for agenda`)
+    console.log(`Found ${agendaItems.length} total items for agenda`)
 
     const opts = {
       ...repo,
@@ -172,7 +116,7 @@ const pkg = require('./package.json')
       createWithin,
       agendaLabel,
       meetingLink,
-      agendaIssues,
+      agendaIssues: agendaItems,
       issueTitle: titleTemplate
     }
